\name{plot_posteriors}
\alias{plot_posteriors}
\title{Plots posteriors for SeqBBSData}
\usage{
  plot_posteriors(seqbbs_data,
    xlab = "Genomic Position, 100kb",
    ylab = "Posterior Probs, all windows", ...)
}
\arguments{
  \item{seqbbs}{SeqBBSData object to work on.  Created
  using seqbbs function.}

  \item{xlab}{x axis label}

  \item{ylab}{y axis label}

  \item{...}{other options passed into plot function}
}
\description{
  Plots posteriors for SeqBBSData
}

